Updated retirement and unit assignments logic in RetirementDefectionDialog
- Added ‘TechConstants’ import utility
- Revisions made to ‘unitAssignmentsComplete’ method to manage edge outcomes
- Updated conditions under which the 'payout' is computed
- Updated the unit assignment enablement logic under 'enableAddRemoveButtons' method
- Modified PersonnelStatus to 'DEFECTED' and 'DESERTED' based on conditions
- Included additional personnel statuses 'POACHED' and 'BURNED_OUT'
- Added logic to handle null 'originalUnit' in 'setOriginalUnit' method
- Adjusted 'bonusMultiplier' calculation to use double instead of int in 'getPayoutOrBonusValue'
- Image size for person’s portrait and force icon in Retirement Table adjusted to 40px from 54px
- Updated 'CustomizePersonDialog' to handle null 'OriginalUnit' selection

+++ MekHQ/src/mekhq/campaign/personnel/turnoverAndRetention/RetirementDefectionTracker.java
@@ -923,14 +923,14 @@ public class RetirementDefectionTracker {
      * @return The amount in C-bills required to get a bonus to the Employee Turnover roll
      */
     public static Money getPayoutOrBonusValue(final Campaign campaign, Person person) {
-        int bonusMultiplier = campaign.getCampaignOptions().getPayoutRateEnlisted();
+        double bonusMultiplier = campaign.getCampaignOptions().getPayoutRateEnlisted();
 
         if (person.getRank().isOfficer()) {
             bonusMultiplier = campaign.getCampaignOptions().getPayoutRateOfficer();
         }
 
         if (campaign.getCampaignOptions().isUsePayoutServiceBonus()) {
-            bonusMultiplier += person.getYearsInService(campaign) * (campaign.getCampaignOptions().getPayoutServiceBonusRate() / 100);
+            bonusMultiplier += person.getYearsInService(campaign) * ((double) campaign.getCampaignOptions().getPayoutServiceBonusRate() / 100);
         }
 
         return person.getSalary(campaign).multipliedBy(bonusMultiplier);
